requirements.txt
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
psutil
pytest
behave